agents/postgres/langchain/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40
langchain-core>=0.1.0
langchain-ollama>=0.1.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.0.0
requests>=2.25.0
fastapi>=0.100.0
uvicorn>=0.20.0
# MCP dependencies
mcp>=1.0.0
asyncio-mqtt>=0.16.0
websockets>=11.0.0
